pyproject.toml
[project]
name = "ai-agent-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.14",
    "python-dotenv>=1.2.2",
]
